Specific now works with Clojure 1.10, 1.9, and 1.8.
Unfortunately, I couldn't find a way to confirm that a spec specifies
a return value spec. The newer versions of clojure.spec default to
clojure.core/any? so checking for existence of a return spec no longer
works.

You could check to see if the return spec is any? but that could be
the actual return spec so I didn't want to enforce that.

This commit also adds a few different profiles for ease of testing
with different Clojure versions and bumps the lein-test-refresh dependency.